Rename rooendpoint
